+++ src/js_sync.hpp
@@ -64,7 +64,11 @@ inline realm::SyncManager& syncManagerShared(typename T::Context &ctx) {
             user_agent_binding_info = js::Value<T>::validated_to_string(ctx, result);
         }
         ensure_directory_exists_for_file(default_realm_file_directory());
-        SyncManager::shared().configure(default_realm_file_directory(), SyncManager::MetadataMode::NoEncryption, user_agent_binding_info);
+        SyncClientConfig client_config;
+        client_config.base_file_path = default_realm_file_directory();
+        client_config.metadata_mode = SyncManager::MetadataMode::NoEncryption;
+        client_config.user_agent_binding_info = user_agent_binding_info;
+        SyncManager::shared().configure(client_config);
     });
     return SyncManager::shared();
 }
@@ -1009,7 +1013,11 @@ void SyncClass<T>::initialize_sync_manager(ContextType ctx, ObjectType this_obje
     args.validate_count(1);
     std::string user_agent_binding_info = Value::validated_to_string(ctx, args[0]);
     ensure_directory_exists_for_file(default_realm_file_directory());
-    SyncManager::shared().configure(default_realm_file_directory(), SyncManager::MetadataMode::NoEncryption, user_agent_binding_info);
+    SyncClientConfig client_config;
+    client_config.base_file_path = default_realm_file_directory();
+    client_config.metadata_mode = SyncManager::MetadataMode::NoEncryption;
+    client_config.user_agent_binding_info = user_agent_binding_info;
+    SyncManager::shared().configure(client_config);
 }
 
 template<typename T>
@@ -1384,7 +1392,7 @@ void SyncClass<T>::local_listener_realms(ContextType ctx, ObjectType this_object
     std::vector<std::string> local_realms;
     for (auto& obj : table) {
         auto virtual_path = obj.get<StringData>(path_col_key);
-        GlobalKey id = obj.get_object_id();
+        auto id = obj.get_object_id();
         std::string file_path = util::format("%1/realms%2/%3.realm", local_root_dir, virtual_path, id.to_string());
 
         // filter out Realms not present locally