+++ tests/js/notifier-tests.ts
@@ -28,7 +28,6 @@ const ipcSchema = [{
         path: 'string',
         insertions: 'int[]',
         deletions: 'int[]',
-        modifications: 'int[]',
         newModifications: 'int[]',
         oldModifications: 'int[]'
     }
@@ -52,7 +51,7 @@ function unexpectedNotificationError(changes) {
     let err = `unexpected notification for ${changes.path} (${Object.keys(changes.changes).length}):\n`;
     for (const objName in changes.changes) {
         const change = changes.changes[objName];
-        err += ` - ${objName}: {${change.insertions}; ${change.deletions}; ${change.modifications}}\n`
+        err += ` - ${objName}: {${change.insertions}; ${change.deletions}; ${change.newModifications}; ${change.oldModifications}}\n`
     }
     return err;
 }
@@ -155,18 +154,24 @@ function availablePromise(path, action) {
     });
 }
 
+function checkArrayEqualityUnordered(array1, array2) {
+    expect(array1.length).toEqual(array2.length);
+    expect(array1.sort()).toEqual(array2.sort());
+}
+
 function notificationPromise(regex, action, notification) {
     regex = `${notificationFilterPrefix}${regex}`;
+    const propertyToCheck = 'int';
     return changeObjectPromise(action, (changes) => {
         expect(changes.path).toMatch(regex);
         expect(Object.keys(changes.changes)).toEqual(Object.keys(notification));
         for (const objectType in notification) {
             const expected = notification[objectType];
             const actual = changes.changes[objectType];
-            expect(actual.insertions).toEqual(expected.insertions || []);
-            expect(actual.deletions).toEqual(expected.deletions || []);
-            expect(actual.modifications).toEqual(expected.modifications || []);
-            expect(actual.oldModifications).toEqual(expected.modifications || []);
+            checkArrayEqualityUnordered(actual.insertions.map(obj => obj[propertyToCheck]), expected.insertions || []);
+            checkArrayEqualityUnordered(actual.deletions.map(obj => obj[propertyToCheck]), expected.deletions || []);
+            checkArrayEqualityUnordered(actual.newModifications.map(obj => obj[propertyToCheck]), expected.newModifications || []);
+            checkArrayEqualityUnordered(actual.oldModifications.map(obj => obj[propertyToCheck]), expected.oldModifications || []);
         }
     });
 }
@@ -231,7 +236,7 @@ describe('Notifier', () => {
         const realm = await rosController.createRealm('test', userRealmSchema);
         await notificationPromise('test',
             () => realm.write(() => realm.create('IntObject', [0])),
-            { IntObject: { insertions: [512] } });
+            { IntObject: { insertions: [0] } });
         realm.close();
 
         // Verify that the GN's copy of the Realm is in the expected place
@@ -284,7 +289,7 @@ describe('Notifier', () => {
         const [callback, realm] = await createRealmAndChangeListener();
         await notificationPromise('test',
             () => realm.write(() => realm.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
         realm.close();
     });
 
@@ -293,7 +298,7 @@ describe('Notifier', () => {
             const [callback, realm] = await createRealmAndChangeListener();
             await notificationPromise('test',
                 () => realm.write(() => realm.create('IntObject', [0])),
-                { IntObject: { insertions: [ 512 ] } });
+                { IntObject: { insertions: [ 0 ] } });
             realm.close();
 
             await Realm.Sync.removeAllListeners();
@@ -316,7 +321,7 @@ describe('Notifier', () => {
             const [callback, realm] = await createRealmAndChangeListener();
             await notificationPromise('test',
                 () => realm.write(() => realm.create('IntObject', [0])),
-                { IntObject: { insertions: [ 512 ] } });
+                { IntObject: { insertions: [ 0 ] } });
             realm.close();
             await Realm.Sync.removeAllListeners();
 
@@ -341,7 +346,7 @@ describe('Notifier', () => {
             const [callback, realm] = await createRealmAndChangeListener();
             await notificationPromise('test',
                 () => realm.write(() => realm.create('IntObject', [0])),
-                { IntObject: { insertions: [ 512 ] } });
+                { IntObject: { insertions: [ 0 ] } });
             realm.close();
             await Realm.Sync.removeAllListeners();
 
@@ -358,12 +363,12 @@ describe('Notifier', () => {
                 realm.create('IntObject', [0]);
                 realm.create('IntObject', [1]);
             }),
-            { IntObject: { insertions: [512, 513] } });
+            { IntObject: { insertions: [0, 1] } });
         await notificationPromise('test', () => realm.write(() => {
                 realm.objects('IntObject')[0].int = 2;
                 realm.delete(realm.objects('IntObject')[1]);
             }),
-            { IntObject: { deletions: [513], modifications: [512] } });
+            { IntObject: { deletions: [1], newModifications: [2], oldModifications: [0] } });
         realm.close();
     });
 
@@ -374,7 +379,7 @@ describe('Notifier', () => {
         await addChangeListener('.*', rosController);
         await notificationPromise('test',
             () => realm.write(() => realm.create('IntObject', [0])),
-            { IntObject: { insertions: [512] } });
+            { IntObject: { insertions: [0] } });
 
         // Remove the listener then add a new one and verify it only gets notified
         // for the new change
@@ -382,7 +387,7 @@ describe('Notifier', () => {
         await addChangeListener('.*', rosController);
         await notificationPromise('test',
             () => realm.write(() => realm.create('IntObject', [0])),
-            { IntObject: { insertions: [513] } });
+            { IntObject: { insertions: [0] } });
         realm.close();
     }, LONG_TIMEOUT);
 
@@ -395,10 +400,10 @@ describe('Notifier', () => {
 
         await notificationPromise('test',
             () => realm1.write(() => realm1.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
         await notificationPromise('test',
             () => realm2.write(() => realm2.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
         realm1.close();
         realm2.close();
     });
@@ -407,7 +412,7 @@ describe('Notifier', () => {
         const [callback, realm] = await createRealmAndChangeListener();
         await notificationPromise('test',
             () => realm.write(() => realm.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
 
         Realm.Sync.removeListener(`${notificationFilterPrefix}.*`, 'change', callback);
         await expectNoNotificationPromise(() =>
@@ -419,7 +424,7 @@ describe('Notifier', () => {
         const [callback, realm] = await createRealmAndChangeListener();
         await notificationPromise('test',
             () => realm.write(() => realm.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
 
         Realm.Sync.removeAllListeners();
         await expectNoNotificationPromise(() =>
@@ -434,7 +439,7 @@ describe('Notifier', () => {
                 realm.create('IntObject', [1]);
                 realm.create('IntObject', [2]);
             }),
-            { IntObject: { insertions: [512, 513] } });
+            { IntObject: { insertions: [1, 2] } });
         await changeObjectPromise(
             () => realm.write(() => realm.objects('IntObject')[0].int = 3),
             (changes) => {
@@ -454,7 +459,7 @@ describe('Notifier', () => {
         ]);
         await notificationPromise('test',
             () => realm1.write(() => realm1.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
         await expectNoNotificationPromise(() =>
             realm2.write(() => realm2.create('IntObject', [0])));
         realm1.close();
@@ -470,12 +475,12 @@ describe('Notifier', () => {
         await addChangeListener('.*test1', rosController);
         await notificationPromise('.*test1',
             () => realm1.write(() => realm1.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
 
         await addChangeListener('.*test2', rosController);
         await notificationPromise('.*test2',
             () => realm2.write(() => realm2.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
         realm1.close();
         realm2.close();
     });
@@ -519,7 +524,7 @@ describe('Notifier', () => {
         await addChangeListener('.*test2', rosController);
         await notificationPromise('.*test2',
             () => realm2.write(() => realm2.create('IntObject', [0])),
-            { IntObject: { insertions: [ 512 ] } });
+            { IntObject: { insertions: [ 0 ] } });
         await addAvailableListener('.*test.', rosController, ['test1', 'test2', 'test3']);
         realm1.close();
         realm2.close();
@@ -861,13 +866,14 @@ describe('Multi-process Notifier', () => {
         }
     }
 
-    function expectEvent(event, type, path, insertions=[], deletions=[], modifications=[]) {
+    function expectEvent(event, type, path, insertions=[], deletions=[], oldModifications=[], newModifications=[]) {
         expect(event.type).toEqual(type);
         expect(event.path).toEqual(`/${currentTestName}/${path}`);
+
         expectArrayishEqual(event.insertions, insertions);
         expectArrayishEqual(event.deletions, deletions);
-        expectArrayishEqual(event.modifications, modifications);
-        expectArrayishEqual(event.oldModifications, modifications);
+        expectArrayishEqual(event.oldModifications, oldModifications);
+        expectArrayishEqual(event.newModifications, newModifications);
     }
 
     // expect an event for each of the realms at the paths present in events
@@ -879,7 +885,7 @@ describe('Multi-process Notifier', () => {
             const expected = events[key];
             expect(expected).toBeDefined();
             delete events[key];
-            expectEvent(actual, expected[0], key, expected[1], expected[1], expected[2]);
+            expectEvent(actual, expected[0], key, expected[1], expected[2], expected[3], expected[4]);
         }
     }
 
@@ -926,17 +932,17 @@ describe('Multi-process Notifier', () => {
         await expectNotifications('test', async (next) => {
             expectEvent(await next(), 'available', 'test');
 
-            realm.write(() => realm.create('IntObject', [0]));
-            expectEvent(await next(), 'change', 'test', [512]);
+            realm.write(() => realm.create('IntObject', [55]));
+            expectEvent(await next(), 'change', 'test', [55]);
 
-            realm.write(() => realm.create('IntObject', [0]));
-            expectEvent(await next(), 'change', 'test', [513]);
+            realm.write(() => realm.create('IntObject', [77]));
+            expectEvent(await next(), 'change', 'test', [77]);
 
-            realm.write(() => { realm.objects('IntObject')[1].int = 2; });
-            expectEvent(await next(), 'change', 'test', [], [], [513]);
+            realm.write(() => { realm.objects('IntObject').filtered('int = 77')[0].int = 2; });
+            expectEvent(await next(), 'change', 'test', [], [], [77], [2]);
 
-            realm.write(() => { realm.delete(realm.objects('IntObject')[0]); });
-            expectEvent(await next(), 'change', 'test', [], [512], []);
+            realm.write(() => { realm.delete(realm.objects('IntObject').filtered('int == 55')[0]); });
+            expectEvent(await next(), 'change', 'test', [], [55], [], []);
         });
         realm.close();
     });